+++ esphome/components/bluetooth_sig_mesh/esp32.h
@@ -16,7 +16,7 @@ class ESP32BluetoothSIGMesh : public BluetoothSIGMesh {
   void loop() override;
 
   void process_mesh_pdu(const uint8_t *data, size_t len) override;
-  void send_mesh_pdu(uint16_t dst, uint16_t app_idx, uint16_t opcode, const uint8_t *payload, size_t len) override;
+  void send_mesh_pdu(uint16_t dst, uint16_t app_idx, uint32_t opcode, const uint8_t *payload, size_t len) override;
   void send_proxy_data_out_notification(const uint8_t *data, size_t len) override;
   void transmit_last_outgoing_frame() override;
