+++ esphome/components/bluetooth_sig_mesh/rp2040.cpp
@@ -61,7 +61,7 @@ void RP2040BluetoothSIGMesh::transmit_last_outgoing_frame() {
   }
 }
 
-void RP2040BluetoothSIGMesh::send_mesh_pdu(uint16_t dst, uint16_t app_idx, uint16_t opcode, const uint8_t *payload,
+void RP2040BluetoothSIGMesh::send_mesh_pdu(uint16_t dst, uint16_t app_idx, uint32_t opcode, const uint8_t *payload,
                                            size_t len) {
   BluetoothSIGMesh::send_mesh_pdu(dst, app_idx, opcode, payload, len);
   this->transmit_last_outgoing_frame();
